+++ resources/views/obat/obat.blade.php
@@ -38,7 +38,7 @@
                     <td data-title="Aksi">
                         <a href="tambah_stok/{{$o->kode}}" class="btn btn-primary"><i class="fa-solid fa-plus"></i> Stok</a>
                         <a href="edit_obat/{{$o->kode}}" class="btn btn-warning"><i class="fa-solid fa-pen-to-square"></i> Edit</a>
-                        <a href="hapus_obat/{{$o->kode}}" class="btn btn-danger"><i class="fa-solid fa-trash"></i> hapus</a>
+                        <a href="hapus_obat/{{$o->kode}}" class="btn btn-danger" onclick="javascript: return confirm('Konfirmasi data akan dihapus');"><i class="fa-solid fa-trash"></i> hapus</a>
                     </td>
                 </tr>
             @endforeach
